Add files via upload
"Worked on my first game project—a Canvas-based Fighting Game—which was an amazing experience! This platform game involved extensive use of JavaScript, along with HTML and CSS, and taught me a lot about game development fundamentals. The tutorial was incredibly helpful, especially for learning OOP concepts and new JavaScript techniques. It was a fantastic introduction to building games!"